refactor(crypto): rename crypto to seed
